SMS-226 BACKEND - updating grades + changed grade field to Double

+++ api/src/main/java/com/sms/grades/GradeDTO.java
@@ -24,7 +24,7 @@ public interface GradeDTO {
 
     String getStudentId();
 
-    Integer getGrade();
+    Double getGrade();
 
     Optional<String> getDescription();
